Add placeholder network, refactor actions into serializable objects
Created an EFTAIOSServer class, which currently has no real
functionality but will help set up the networking functionality. Also
created a ChatMessage object, which compounds text, the player speaking,
and a time stamp to create uniquely identifiable chat objects for when
the game client pulls from the server. Created framework for actions &
action history, but the functionality is not yet in place