Improve GUI, rebase Server framework, add server connection panel
The main game window GUI now has stripes behind it, to provide a bit of
contrast and to resemble the real game map. I've remodeled the base
networking code to instead host a multi-threaded server. I've had to
break a lot of the functionality, because I'm overhauling the pittance
of functionality to the network based system. Chat no longer works, as
the previous implementation kept a local game engine object and used it
to query the state. Now that is not possible, because the server is what
stores the game object. Most listeners and interfaces to do with the
GameEngine have been removed. A server lobby interface has been created
which "allows" hosting and joining servers. However, the functionality
is not entirely there, so it currently serves as a transition between
the two scenes. Eventually this will lead to an intermittent game lobby
window, but that will be soon to come.